BRMS-117: refactor into seperate packages
- moving common functionality into own packages
- adding unit tests for UTF-8 characters
