+++ app/uk/gov/hmrc/brm/controllers/MatchingController.scala
@@ -16,11 +16,8 @@
 
 package uk.gov.hmrc.brm.controllers
 
-import java.time.LocalDate
-
-import org.joda.time.format.DateTimeFormat
 import play.api.Logger
-import play.api.libs.json.{JsObject, JsArray, Json}
+import play.api.libs.json.{JsArray, JsObject}
 import play.api.mvc.{Action, Request, Result}
 import uk.gov.hmrc.brm.connectors._
 import uk.gov.hmrc.brm.utils.BrmLogger._
@@ -29,7 +26,6 @@ import uk.gov.hmrc.play.http.{Upstream4xxResponse, Upstream5xxResponse}
 import uk.gov.hmrc.play.microservice.controller.BaseController
 
 import scala.concurrent.Future
-import scala.util.{Failure, Success, Try}
 
 
 object MatchingController extends MatchingController {
@@ -78,7 +74,7 @@ trait MatchingController extends BaseController {
 
   def success(method: String): PartialFunction[BirthResponse, Future[Result]] = {
     case BirthSuccessResponse(js) =>
-      val count = if(js.isInstanceOf[JsArray]) js.as[JsArray].value.length else js.asOpt[JsObject].fold(0)(x => 1)
+      val count = if(js.isInstanceOf[JsArray]) js.as[JsArray].value.length else 1
       info(CLASS_NAME, s"$method", s"success: $count record(s) found")
       respond(Ok(js))
   }
@@ -86,20 +82,19 @@ trait MatchingController extends BaseController {
   def reference(reference: String) = Action.async {
     implicit request =>
       setKey(request)
-
       Logger.debug(s"connector: ${groConnector.get(reference)}")
-
       groConnector.get(reference).flatMap[Result](
         handleException("getReference", reference)
         orElse success("getReference")
       )
   }
 
-  def details(firstName: String, lastName: String, dateOfBirth: String) = Action.async {
+  def details(forenames: String, lastname: String, dateofbirth: String) = Action.async {
     implicit request =>
       setKey(request)
-      groConnector.get(firstName, lastName, dateOfBirth).flatMap[Result](
-        handleException("getDetails", dateOfBirth)
+      debug(CLASS_NAME, "details", s"firstName $forenames, lastName: $lastname, dateOfBirth: $dateofbirth")
+      groConnector.get(forenames, lastname, dateofbirth).flatMap[Result](
+        handleException("getDetails", dateofbirth)
         orElse success("getDetails")
       )
   }
