+++ conf/resources/gro/2006-11-12_smith_adam-utf-8.json
@@ -1,0 +1,127 @@
+[
+  {
+    "location": {
+      "name": "Ledbury, Ledbury, County of Herefordshire",
+      "administrativeArea": "County of Herefordshire",
+      "subDistrict": "Ledbury",
+      "nameQualifier": "generated by concatenating registrationDistrict, subDistrict and administrativeArea",
+      "registrationDistrict": "Ledbury"
+    },
+    "registrarSignature": "Asdfasdf",
+    "subjects": {
+      "child": {
+        "originalName": {},
+        "name": {
+          "givenName": "Adàm TËST",
+          "surname": "SMÏTH",
+          "fullName": "Adàm TËST SMÏTH",
+          "qualifier": "generated from prepending forename to surname"
+        },
+        "dateOfBirth": "2006-11-12",
+        "sex": "Male",
+        "birthplace": "24 Church Road Ledbury"
+      },
+      "father": {
+        "name": {
+          "givenName": "Asdf",
+          "surname": "ASDF",
+          "fullName": "Asdf ASDF",
+          "qualifier": "generated from prepending forename to surname"
+        },
+        "birthplace": "Asdf",
+        "occupation": "Asdfasdf"
+      },
+      "mother": {
+        "name": {
+          "givenName": "Joan",
+          "surname": "SMITH",
+          "fullName": "Joan SMITH",
+          "qualifier": "generated from prepending forename to surname"
+        },
+        "birthplace": "Hereford",
+        "occupation": "Housewife",
+        "usualAddress": "24 Church Road Ledbury"
+      },
+      "informant": {
+        "name": {},
+        "usualAddress": "24 Church Road Ledbury",
+        "qualification": "Mother",
+        "signature": "J Smith"
+      }
+    },
+    "systemNumber": 500035710,
+    "id": 2135,
+    "date": "2008-11-12",
+    "status": {
+      "potentiallyFictitiousBirth": false,
+      "correction": "None",
+      "cancelled": false,
+      "blockedRegistration": false,
+      "marginalNote": "None",
+      "reRegistered": "None"
+    },
+    "previousRegistration": {}
+  },{
+  "location": {
+    "name": "Ledbury, Ledbury, County of Herefordshire",
+    "administrativeArea": "County of Herefordshire",
+    "subDistrict": "Ledbury",
+    "nameQualifier": "generated by concatenating registrationDistrict, subDistrict and administrativeArea",
+    "registrationDistrict": "Ledbury"
+  },
+  "registrarSignature": "Asdfasdf",
+  "subjects": {
+    "child": {
+      "originalName": {},
+      "name": {
+        "givenName": "Adam TEST",
+        "surname": "SMITH",
+        "fullName": "Adam TEST SMITH",
+        "qualifier": "generated from prepending forename to surname"
+      },
+      "dateOfBirth": "2006-11-12",
+      "sex": "Male",
+      "birthplace": "24 Church Road Ledbury"
+    },
+    "father": {
+      "name": {
+        "givenName": "Asdf",
+        "surname": "ASDF",
+        "fullName": "Asdf ASDF",
+        "qualifier": "generated from prepending forename to surname"
+      },
+      "birthplace": "Asdf",
+      "occupation": "Asdfasdf"
+    },
+    "mother": {
+      "name": {
+        "givenName": "Joan",
+        "surname": "SMITH",
+        "fullName": "Joan SMITH",
+        "qualifier": "generated from prepending forename to surname"
+      },
+      "birthplace": "Hereford",
+      "occupation": "Housewife",
+      "usualAddress": "24 Church Road Ledbury"
+    },
+    "informant": {
+      "name": {},
+      "usualAddress": "24 Church Road Ledbury",
+      "qualification": "Mother",
+      "signature": "J Smith"
+    }
+  },
+  "systemNumber": 500035710,
+  "id": 2135,
+  "date": "2008-11-12",
+  "status": {
+    "potentiallyFictitiousBirth": false,
+    "correction": "None",
+    "cancelled": false,
+    "blockedRegistration": false,
+    "marginalNote": "None",
+    "reRegistered": "None"
+  },
+  "previousRegistration": {}
+}
+]
