+++ test/uk/gov/hmrc/brm/metrics/MetricsSpec.scala
@@ -16,8 +16,6 @@
 
 package uk.gov.hmrc.brm.metrics
 
-import java.util.concurrent.TimeUnit
-
 import com.kenshoo.play.metrics.MetricsRegistry
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
